auto_data_gen_val/requirements.txt
adal==1.2.7
aiohttp==3.8.5
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.29.2
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
cmake==3.27.1
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dataclasses-json==0.5.14
decorator==5.1.1
docx2txt==0.8
dominate==2.8.0
executing==1.2.0
filelock==3.12.2
filetype==1.2.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gevent==23.7.0
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipython==8.12.2
ipywidgets==8.1.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.4
langchain==0.0.257
langsmith==0.0.19
lit==16.0.6
llama-index==0.7.21
llvmlite==0.40.1
load-dotenv==0.1.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
microsoft-bing-autosuggest==1.0.0
microsoft-bing-customimagesearch==1.0.0
microsoft-bing-customwebsearch==1.0.0
microsoft-bing-entitysearch==1.0.0
microsoft-bing-imagesearch==1.0.0
microsoft-bing-newssearch==1.0.0
microsoft-bing-spellcheck==1.0.0
microsoft-bing-videosearch==1.0.0
microsoft-bing-visualsearch==1.0.0
microsoft-bing-websearch==1.0.0
more-itertools==10.1.0
mpmath==1.3.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
plotly==5.15.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.3.0
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.3
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.9.1
urllib3==1.26.16
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.3.6
widgetsnbextension==4.0.8
xgboost==1.7.6
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
